pyproject.toml
[tool.poetry]
name = "nux"
version = "0.1.0"
description = ""
authors = ["i3cheese <i3cheese@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
ConfigArgParse = "^1.5.3"
SQLAlchemy = "^1.4.37"
uvicorn = "^0.17.6"
psycopg2 = "^2.9.3"
pydantic = "^1.9.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
alembic = "^1.8.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
yarl = "^1.7.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
SQLAlchemy-Utils = "^0.38.2"
requests = "^2.28.0"

[tool.poetry.scripts]
run = "nux:run_app"
create_all = "nux.database:create_all"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[pycodestyle]
max_line_length = 120
